+++ drivers/counter/counter_ace_v1x_art.c
@@ -8,7 +8,6 @@
 #include <zephyr/devicetree.h>
 #include <zephyr/drivers/counter.h>
 #include <soc.h>
-#include <ace_v1x-regs.h>
 #include <counter/counter_ace_v1x_art_regs.h>
 
 static struct k_spinlock lock;
