+++ drivers/counter/counter_ace_v1x_rtc.c
@@ -8,7 +8,6 @@
 #include <zephyr/devicetree.h>
 #include <zephyr/drivers/counter.h>
 #include <soc.h>
-#include <ace_v1x-regs.h>
 #include <counter/counter_ace_v1x_rtc_regs.h>
 
 static int counter_ace_v1x_rtc_get_value(const struct device *dev,
