+++ drivers/mm/mm_drv_intel_adsp_mtl_tlb.c
@@ -34,7 +34,6 @@
 
 #include <soc.h>
 #include <adsp_memory.h>
-#include <ace_v1x-regs.h>
 
 #include "mm_drv_common.h"
