+++ soc/xtensa/intel_adsp/common/ipc.c
@@ -157,8 +157,6 @@ bool intel_adsp_ipc_send_message_sync(const struct device *dev,
 #if DT_NODE_EXISTS(INTEL_ADSP_IPC_HOST_DTNODE)
 
 #if defined(CONFIG_SOC_SERIES_INTEL_ACE)
-#include <ace_v1x-regs.h>
-
 static inline void ace_ipc_intc_unmask(void)
 {
 	for (int i = 0; i < CONFIG_MP_NUM_CPUS; i++) {
